feat: Launch child
