+++ lib/node-inspect.js
@@ -20,261 +20,27 @@
  * IN THE SOFTWARE.
  */
 'use strict';
-const { EventEmitter } = require('events');
-const http = require('http');
-const util = require('util');
+const spawn = require('child_process').spawn;
+const Repl = require('repl');
 
-const debuglog = util.debuglog('inspect');
+const ProtocolClient = require('./internal/inspect-protocol');
 
 exports.port = 9229;
 
-const ProtocolClient = (() => {
-  const kOpCodeText = 0x1;
-  const kOpCodeClose = 0x8;
+function throwUnexpectedError(error) {
+  process.nextTick(() => { throw error; });
+}
 
-  const kFinalBit = 0x80;
-  const kReserved1Bit = 0x40;
-  const kReserved2Bit = 0x20;
-  const kReserved3Bit = 0x10;
-  const kOpCodeMask = 0xF;
-  const kMaskBit = 0x80;
-  const kPayloadLengthMask = 0x7F;
-
-  const kMaxSingleBytePayloadLength = 125;
-  const kMaxTwoBytePayloadLength = 0xFFFF;
-  const kTwoBytePayloadLengthField = 126;
-  const kEightBytePayloadLengthField = 127;
-  const kMaskingKeyWidthInBytes = 4;
-
-  function ignoreError() {}
-
-  function isEmpty(obj) {
-    return Object.keys(obj).length === 0;
-  }
-
-  function unpackError({ code, message, data }) {
-    const err = new Error(`${message} - ${data}`);
-    err.code = code;
-    Error.captureStackTrace(err, unpackError);
-    return err;
-  }
-
-  function encodeFrameHybi17(payload) {
-    const dataLength = payload.length;
-
-    let singleByteLength;
-    let additionalLength;
-    if (dataLength > kMaxTwoBytePayloadLength) {
-      singleByteLength = kEightBytePayloadLengthField;
-      additionalLength = new Buffer(8);
-      let remaining = dataLength;
-      for (let i = 0; i < 8; ++i) {
-        additionalLength[7 - i] = remaining & 0xFF;
-        remaining >>= 8;
-      }
-    } else if (dataLength > kMaxSingleBytePayloadLength) {
-      singleByteLength = kTwoBytePayloadLengthField;
-      additionalLength = new Buffer(2);
-      additionalLength[0] = (dataLength & 0xFF00) >> 8;
-      additionalLength[1] = dataLength & 0xFF;
-    } else {
-      additionalLength = new Buffer(0);
-      singleByteLength = dataLength;
-    }
-
-    const header = new Buffer([
-      kFinalBit | kOpCodeText,
-      kMaskBit | singleByteLength,
-    ]);
-
-    const mask = new Buffer(4);
-    const masked = new Buffer(dataLength);
-    for (let i = 0; i < dataLength; ++i) {
-      masked[i] = payload[i] ^ mask[i % kMaskingKeyWidthInBytes];
-    }
-
-    return Buffer.concat([header, additionalLength, mask, masked]);
-  }
-
-  function decodeFrameHybi17(data) {
-    const dataAvailable = data.length;
-    const notComplete = { closed: false, payload: null, rest: data };
-    let payloadOffset = 2;
-    if ((dataAvailable - payloadOffset) < 0) return notComplete;
-
-    const firstByte = data[0];
-    const secondByte = data[1];
-
-    const final = (firstByte & kFinalBit) !== 0;
-    const reserved1 = (firstByte & kReserved1Bit) !== 0;
-    const reserved2 = (firstByte & kReserved2Bit) !== 0;
-    const reserved3 = (firstByte & kReserved3Bit) !== 0;
-    const opCode = firstByte & kOpCodeMask;
-    const masked = (secondByte & kMaskBit) !== 0;
-    const compressed = reserved1;
-    if (compressed) {
-      throw new Error('Compressed frames not supported');
-    }
-    if (!final || reserved2 || reserved3) {
-      throw new Error('Only compression extension is supported');
-    }
-
-    if (masked) {
-      throw new Error('Masked server frame - not supported');
-    }
-
-    let closed = false;
-    switch (opCode) {
-      case kOpCodeClose:
-        closed = true;
-        break;
-      case kOpCodeText:
-        break;
-      default:
-        throw new Error(`Unsupported op code ${opCode}`);
-    }
-
-    let payloadLength = secondByte & kPayloadLengthMask;
-    switch (payloadLength) {
-      case kTwoBytePayloadLengthField:
-        payloadOffset += 2;
-        payloadLength = (data[2] << 8) + data[3];
-        break;
-
-      case kEightBytePayloadLengthField:
-        payloadOffset += 8;
-        payloadLength = 0;
-        for (let i = 0; i < 8; ++i) {
-          payloadLength <<= 8;
-          payloadLength |= data[2 + i];
-        }
-        break;
-
-      default:
-        // Nothing. We already have the right size.
-    }
-    if ((dataAvailable - payloadOffset) < 0) return notComplete;
-
-    const payloadEnd = payloadOffset + payloadLength;
-    return {
-      payload: data.slice(payloadOffset, payloadEnd),
-      rest: data.slice(payloadEnd),
-      closed,
-    };
-  }
-
-  class Client extends EventEmitter {
-    constructor() {
-      super();
-      this._http = null;
-      this._open = false;
-      this._lastId = 0;
-      this._socket = null;
-      this._pending = {};
-      this._unprocessed = new Buffer(0);
-
-      this.handleChunk = this._handleChunk.bind(this);
-    }
-
-    _handleChunk(chunk) {
-      this._unprocessed = Buffer.concat([this._unprocessed, chunk]);
-
-      while (this._unprocessed.length > 2) {
-        const { payload, rest } = decodeFrameHybi17(this._unprocessed);
-        this._unprocessed = rest;
-        if (payload === null) break;
-
-        debuglog('< %s', payload);
-        const { id, method, params, result, error } = JSON.parse(payload.toString());
-        if (id) {
-          const handler = this._pending[id];
-          if (handler) {
-            delete this._pending[id];
-            handler(error, result);
-          }
-        } else if (method) {
-          this.emit('debugEvent', method, params);
-          this.emit(method, params);
-        } else {
-          throw new Error(`Unsupported response: ${payload.toString()}`);
-        }
-      }
-    }
-
-    destroy() {
-      if (this._http) {
-        this._http.destroy();
-        this._http = null;
-      }
-    }
-
-    callMethod(method, params) {
-      return new Promise((resolve, reject) => {
-        const data = { id: ++this._lastId, method, params };
-        this._pending[data.id] = (error, result) => {
-          if (error) reject(unpackError(error));
-          else resolve(isEmpty(result) ? undefined : result);
-        };
-        const json = JSON.stringify(data);
-        debuglog('> %s', json);
-        this._socket.write(encodeFrameHybi17(new Buffer(json)));
-      });
-    }
-
-    connect(port, host) {
-      const key1 = crypto.randomBytes(16).toString('base64');
-
-      const httpReq = this._http = http.request({
-        host, port,
-        path: '/node',
-        headers: {
-          Connection: 'Upgrade',
-          Upgrade: 'websocket',
-          'Sec-WebSocket-Key': key1,
-        },
-      });
-      httpReq.on('error', e => {
-        this.emit('error', e);
-      });
-      httpReq.on('response', httpRes => {
-        if (httpRes.statusCode >= 400) {
-          httpRes.pipe(process.stderr);
-        } else {
-          httpRes.pipe(process.stderr);
-        }
-      });
-
-      const handshakeListener = (res, socket) => {
-        // TODO: we *could* validate res.headers[sec-websocket-accept]
-
-        this._socket = socket;
-        socket.on('data', this.handleChunk);
-
-        Promise.all([
-          this.callMethod('Log.enable').then(null, ignoreError),
-          this.callMethod('Runtime.enable'),
-          this.callMethod('Page.enable').then(null, ignoreError),
-          this.callMethod('Page.getResourceTree').then(null, ignoreError),
-          this.callMethod('Debugger.enable'),
-          this.callMethod('Debugger.setPauseOnExceptions', { state: 'none' }),
-          this.callMethod('Debugger.setAsyncCallStackDepth', { maxDepth: 0 }),
-          this.callMethod('Profiler.enable'),
-          this.callMethod('Profiler.setSamplingInterval', { interval: 100 }),
-          this.callMethod('Debugger.setBlackboxPatterns', { patterns: [] }),
-          this.callMethod('Runtime.run'),
-        ]).then(() => {
-          this.emit('ready');
-        }, error => {
-          this.emit('error', error);
-        });
-      };
-      httpReq.on('upgrade', handshakeListener);
-      httpReq.end();
-    }
-  }
-
-  return Client;
-})();
+function runScript(script, scriptArgs, inspectPort) {
+  return new Promise((resolve) => {
+    const args = [
+      '--inspect',
+      `--debug-brk=${inspectPort}`,
+    ].concat([script], scriptArgs);
+    const child = spawn(process.execPath, args);
+    resolve(child);
+  });
+}
 
 class NodeInspector {
   constructor(options, stdin, stdout) {
@@ -282,11 +48,84 @@ class NodeInspector {
     this.stdin = stdin;
     this.stdout = stdout;
 
-    this.client = null;
+    this.paused = false;
+    this.child = null;
+
+    this._runScript = options.script ?
+      runScript.bind(null, options.script, options.scriptArgs, options.port) :
+      Promise.resolve.bind(null, null);
+
+    this.client = new ProtocolClient(options.port, options.host);
+
+    // Handle all possible exits
+    process.on('exit', () => this.killChild());
+    process.once('SIGTERM', process.exit.bind(process, 0));
+    process.once('SIGHUP', process.exit.bind(process, 0));
+
+    this.run()
+      .then(() => this._setupRepl())
+      .then(null, throwUnexpectedError);
   }
 
-  connect() {
-    this.client = new ProtocolClient();
+  _setupRepl() {
+    const replOptions = {
+      prompt: 'debug> ',
+      input: this.stdin,
+      output: this.stdout,
+      eval: (code, ctx, file, cb) => this.controlEval(code, ctx, file, cb),
+      useGlobal: false,
+      ignoreUndefined: true,
+    };
+
+    this.repl = Repl.start(replOptions);
+    // Kill child process when main process dies
+    this.repl.on('exit', () => {
+      process.exit(0);
+    });
+  }
+
+  killChild() {
+    this.client.reset();
+    if (this.child) {
+      this.child.kill();
+      this.child = null;
+    }
+  }
+
+  run() {
+    this.killChild();
+    return this._runScript().then((child) => {
+      this.child = child;
+      this.child.stdout.on('data', text => this.childPrint(text));
+      this.child.stderr.on('data', text => this.childPrint(text));
+    });
+  }
+
+  clearline() {
+    if (this.stdout.isTTY) {
+      this.stdout.cursorTo(0);
+      this.stdout.clearLine(1);
+    } else {
+      this.stdout.write('\b');
+    }
+  }
+
+  print(text, oneline = false) {
+    this.clearline();
+    this.stdout.write(oneline ? `${text}\n` : text);
+  }
+
+  childPrint(text) {
+    this.print(
+      text.toString()
+        .split(/\r\n|\r|\n/g)
+        .filter(chunk => !!chunk)
+        .map(chunk => `< ${chunk}`)
+        .join('\n')
+    );
+    if (!this.paused) {
+      this.repl.displayPrompt(true);
+    }
   }
 }
 
@@ -295,20 +134,28 @@ function parseArgv([target, ...args]) {
   let port = exports.port;
   let isRemote = false;
   let script = target;
+  let scriptArgs = args;
 
   const hostMatch = target.match(/^([^:]+):(\d+)$/);
-
+  const portMatch = target.match(/^--port=(\d+)$/);
   if (hostMatch) {
-    // Connecting to remote debugger, e.g. `node-inspect localhost:5858`
+    // Connecting to remote debugger
+    // `node-inspect localhost:9229`
     host = hostMatch[1];
     port = parseInt(hostMatch[2], 10);
     isRemote = true;
     script = null;
+  } else if (portMatch) {
+    // Start debugger on custom port
+    // `node debug --port=8058 app.js`
+    port = parseInt(portMatch[1], 10);
+    script = args[0];
+    scriptArgs = args.slice(1);
   }
 
   return {
     host, port,
-    isRemote, script, args,
+    isRemote, script, scriptArgs,
   };
 }
 
@@ -327,14 +174,16 @@ function start(argv = process.argv.slice(2),
 
   stdin.resume();
 
-  process.on('uncaughtException', e => {
+  function handleUnexpectedError(e) {
     console.error('There was an internal error in node-inspect. ' +
                   'Please report this bug.');
     console.error(e.message);
     console.error(e.stack);
     if (inspector.child) inspector.child.kill();
     process.exit(1);
-  });
+  }
+
+  process.on('uncaughtException', handleUnexpectedError);
   /* eslint-enable no-console */
 }
 exports.start = start;
